# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Custom Prysm endpoints for beacon node operations, validator metrics, and extended functionality.\nThese endpoints extend the standard Ethereum Beacon Node API with Prysm-specific features.",
        "title": "Prysm Beacon Node Custom API",
        "contact": {
            "name": "Prysm Team",
            "url": "https://github.com/OffchainLabs/prysm"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://github.com/OffchainLabs/prysm/blob/develop/LICENSE.md"
        },
        "version": "1.0"
    },
    "host": "localhost:3500",
    "basePath": "/",
    "paths": {
...
        "/prysm/node/trusted_peers": {
            "get": {
                "description": "Returns a list of all trusted peers configured on this beacon node, including their connection state, direction, ENR, and last seen address",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prysm Node"
                ],
                "summary": "List trusted peers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/structs.PeersResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.DefaultJsonError"
                        }
                    }
                }
            },
            "post": {
                "description": "Adds a peer to the beacon node's trusted peer set using its multiaddress. Trusted peers are given priority in peer connections.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Prysm Node"
                ],
                "summary": "Add a trusted peer",
                "parameters": [
                    {
                        "description": "Peer multiaddress (e.g., /ip4/127.0.0.1/tcp/13000/p2p/16Uiu2...)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/structs.AddrRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Peer successfully added to trusted set"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.DefaultJsonError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.DefaultJsonError"
                        }
                    }
                }
            }
        },
        "/prysm/node/trusted_peers/{peer_id}": {
            "delete": {
                "description": "Removes a peer from the beacon node's trusted peer set by peer ID. Does not disconnect the peer if already connected.",
                "tags": [
                    "Prysm Node"
                ],
                "summary": "Remove a trusted peer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Peer ID to remove from trusted set",
                        "name": "peer_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Peer successfully removed from trusted set"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.DefaultJsonError"
                        }
                    }
                }
            }
        },
...
        "/prysm/v1/validators/{state_id}/active_set_changes": {
            "get": {
                "description": "Returns validator set changes for a given epoch including activations, voluntary exits, slashings, and ejections with both public keys and indices",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prysm Validator"
                ],
                "summary": "Get active validator set changes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "State identifier (head, genesis, finalized, justified, slot number, or hex root)",
                        "name": "state_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/structs.ActiveSetChanges"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.DefaultJsonError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.DefaultJsonError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.DefaultJsonError"
                        }
                    }
                }
            }
        },
        "/prysm/v1/validators/{state_id}/participation": {
            "get": {
                "description": "Returns participation rate information for validators at a given state, including voting statistics and balance information for current and previous epochs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Prysm Validator"
                ],
                "summary": "Get validator participation for an epoch",
                "parameters": [
                    {
                        "type": "string",
                        "description": "State identifier (head, genesis, finalized, justified, slot number, or hex root)",
                        "name": "state_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/structs.GetValidatorParticipationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.DefaultJsonError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.DefaultJsonError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.DefaultJsonError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "httputil.DefaultJsonError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "structs.ActiveSetChanges": {
            "type": "object",
            "properties": {
                "activated_indices": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "activated_public_keys": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ejected_indices": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ejected_public_keys": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "epoch": {
                    "type": "string"
                },
                "exited_indices": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "exited_public_keys": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "slashed_indices": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "slashed_public_keys": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "structs.AddrRequest": {
            "type": "object",
            "properties": {
                "addr": {
                    "type": "string"
                }
            }
        },
...
        "structs.GetValidatorParticipationResponse": {
            "type": "object",
            "properties": {
                "epoch": {
                    "type": "string"
                },
                "finalized": {
                    "type": "boolean"
                },
                "participation": {
                    "$ref": "#/definitions/structs.ValidatorParticipation"
                }
            }
        },
...
        "structs.Peer": {
            "type": "object",
            "properties": {
                "direction": {
                    "type": "string"
                },
                "enr": {
                    "type": "string"
                },
                "last_seen_p2p_address": {
                    "type": "string"
                },
                "peer_id": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
        "structs.PeersResponse": {
            "type": "object",
            "properties": {
                "peers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/structs.Peer"
                    }
                }
            }
        },
...
        "structs.ValidatorParticipation": {
            "type": "object",
            "properties": {
                "current_epoch_active_gwei": {
                    "type": "string"
                },
                "current_epoch_attesting_gwei": {
                    "type": "string"
                },
                "current_epoch_target_attesting_gwei": {
                    "type": "string"
                },
                "eligible_ether": {
                    "type": "string"
                },
                "global_participation_rate": {
                    "type": "string"
                },
                "previous_epoch_active_gwei": {
                    "type": "string"
                },
                "previous_epoch_attesting_gwei": {
                    "type": "string"
                },
                "previous_epoch_head_attesting_gwei": {
                    "type": "string"
                },
                "previous_epoch_target_attesting_gwei": {
                    "type": "string"
                },
                "voted_ether": {
                    "type": "string"
                }
            }
        },
...
    }
}